+++ packages/components/barchart/CHANGELOG.md
@@ -1,3 +1,7 @@
+## `@svizzle/barchart` v0.2.0 (next)
+
+- `BarchartV` accepts `keyToColorFn`
+
 ## `@svizzle/barchart` v0.1.0
 
-- added the component
+- added `BarchartV`