+++ packages/components/choropleth/CHANGELOG.md
@@ -1,3 +1,11 @@
-## `@svizzle/choropleth` v0.1.0 (next)
+## `@svizzle/choropleth` v0.2.0 (next)
 
-- added the component
+- `ChoroplethSVG` accepts `keyToColorFn`
+- `ChoroplethDiv` accepts `keyToColorFn`
+- `ChoroplethWorldDiv` accepts `keyToColorFn`
+
+## `@svizzle/choropleth` v0.1.0
+
+- added `ChoroplethSVG`
+- added `ChoroplethDiv`
+- added `ChoroplethWorldDiv`